separation of concerns
